Introduce test users
Now we have an admin and a plain user, "hello".

They are hard-coded in the configuration file. They should be only
development purpose and they should not be in the production code.

To make our life easier, I changed the `email` type to `text` type input
for the login form. Because "admin" and "hello" are not in the email
format.